Fix images saving as base64 (#93)
When saving a post, there is code to extract the base64 encoded images into files and update the post content to point to those newly saved files.  However, this runs after the blog post is saved, so the updates to the blog post's markup aren't persisted.  The files are correctly saved, but the post's markup still includes base64 encoded images.